suts/proshop-api-B648/requirements.txt
psycopg2-binary
fastapi
python-dotenv
uvicorn
sqlalchemy
pydantic
fastapi-sqlalchemy